feat: 몇 가지 출력 기능 추가와 코드 개선
누락된 도전 과제 기능 추가
- 이동 종료 후 총 이동 거리 출력 기능 추가.
- 이동 종료 후 총 이벤트 발생 횟수 출력해주는 기능 추가.
- Tick 10회 카운트 후 Tick 발생을 정지 시킴

코드 개선
- tick 10회 카운트 조건을 Tick()으로 이전해 신호 전달 안정성을 더함.

+++ Source/work_5/Public/MyActor2.h
@@ -15,7 +15,7 @@ private:
 	FVector2D start;
 	int32_t tickCount;
 	int32_t evCnt;
-	int totDist;
+	float totDist;
 
 public:
 	// Sets default values for this actor's properties
